exam_seating_system-main/exam_seating_system-main/requirements.txt
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.6
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
gcloud==0.18.3
googleapis-common-protos==1.63.0
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jwcrypto==1.5.6
MarkupSafe==2.1.5
oauth2client==4.1.3
openpyxl==3.1.2
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-jwt==4.1.0
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==1.26.18
uWSGI==2.0.25.1
Werkzeug==3.0.2
